feat: AGP-MCP integration
